Feat: 재생하기 클릭시 ViewCount+1 추가하고 서비스준비중 메세지출력

+++ animation-backend/src/main/java/com/example/boot/Entity/Animation.java
@@ -26,7 +26,7 @@ public class Animation {
     private String dayOfWeek;
     private double averageRating;
     private String uploaded;
-    private int viewCount;
+    private Long viewCount;
     private String viewed;
     private LocalDateTime viewedTime;
     private String favorite;
